pyproject.toml
[project]
name = "bioagent-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "litellm>=1.74.9.post1",
    "python-dotenv>=1.1.1",
]
